iirc/requirements.txt
torch
torchvision
numpy==1.18.5
Pillow==7.0.0
lmdb==1.0.0
git+https://git@github.com/shagunsodhani/ml-logger.git@976cab1d2524ee72aef8273e6deb25c764883f3e#egg=mllogger[all]
seaborn==0.10.1
pytest==5.4.3
pytest-cov==2.9.0
wandb-offline-sync